pyproject.toml
[tool.poetry]
name = "podomarket"
version = "0.1.0"
description = ""
authors = ["Heo-JunHyeok <junhyeok_1996@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^4.2.3"
django-dotenv = "^1.4.2"
django-allauth = "^0.54.0"
django-widget-tweaks = "^1.4.12"
pillow = "^10.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
